somesite/requirements.txt
asgiref==3.5.2
awsebcli==3.20.3
backports.zoneinfo==0.2.1
black==22.3.0
botocore==1.23.54
cement==2.8.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.3
Django==2.2
djangorestframework==3.13.1
future==0.16.0
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==304
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.1.9
